Avoid the word "session"
RFC 4210 defines the term "session key" to refer to the ticket session
key, which is not impacted by pre-authentication.  Any use of that
term within this document could be very confusing to implementors.
The SPAKE2 draft does not use the word "session," fortunately.

This leaves us with slightly too few unique terms to describe w, K,
and K'[n].  With this commit, the term "shared secret" is used to
describe both w and K, with care to use terms like "shared secret
input w" and "calculated shared secret K" to avoid ambiguity.  The
term "encryption key" is used to describe K'[n].