Permit x and y reuse
The SPAKE2 draft doesn't say anything about reusing x and y. We
believe there is no problem with reuse except for the impact on
forward secrecy (which we should discuss in security considerations;
that will come in a later commit). Add a note about the selection of x
and y values to the SPAKE parameters section, and permit them to be
reused.