Improve RFC 6113 conformance
To permit the mechanism to be used as part of an authentication set,
use the term "initial reply key" to refer to the client long-term key,
and use similar language to explain why a single ETYPE-INFO-ENTRY must
be used when the mechanism is stand-alone.

Add a paragraph to the top of the main description indicating that the
reply key is used, describing which facilities are provided,
summarizing what is true at the end of the mechanism, and stating that
no pa-hint value is required when used in an authentication set.

Remove some unnecessary sentences which assume that the mechanism ends
with an AS-REP. Add a new subsection on mechanism
termination. Describe the reply key change as strengthening rather
than replacement, as we use a derivative of the initial reply key.

In security considerations, specifically talk about unauthenticated
plaintext and how it is protected. Talk about replays to different
KDCs. Talk about the lack of direct KDC authentication.