Expand document conventions section
Add text to the document conventions section referring to RFC 4120,
RFC 3961, and RFC 6113. The RFC 4120 references aren't spelled out
because there would be dozens of them; this is also done in RFC 6113's
conventions section. Explain that "SPAKE" means a variant of
SPAKE2. Use fewer explicit references in the main text.