spake: Fix four issues
Change "SPAKEKey" to "SPAKEkey" for case consistency with
"SPAKEsecret".

Include the group number in the derivation of w so that we don't use
similar or identical secrets with different groups.

Include the group number and enctype number in the derivation of K'[n]
to ensure that there are never PRF+ input string collisions across
different groups or enctypes.

Specify that the KDC must supply ETYPE-INFO2 information to the client
when responding to an optimistic support message.